# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AWS Device Farm Hub API",
    "version": "1.0.0",
    "description": "Complete API for managing AWS Device Farm projects, apps, and test runs",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "tags": [
    {
      "name": "Projects",
      "description": "Device Farm project management"
    },
    {
      "name": "Device Pools",
      "description": "Device pool management"
    },
    {
      "name": "Devices",
      "description": "Available devices"
    },
    {
      "name": "Apps",
      "description": "Application upload and management"
    },
    {
      "name": "Runs",
      "description": "Test run execution and monitoring"
    },
    {
      "name": "Artifacts",
      "description": "Test artifacts and results"
    }
  ],
  "paths": {
    "/api/projects": {
      "get": {
        "tags": ["Projects"],
        "summary": "List all projects",
        "description": "Get all Device Farm projects for the authenticated user",
        "responses": {
          "200": {
            "description": "List of projects",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Project"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "post": {
        "tags": ["Projects"],
        "summary": "Create a new project",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Project created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{id}": {
      "get": {
        "tags": ["Projects"],
        "summary": "Get project by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Project details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["Projects"],
        "summary": "Update project",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Project updated"
          }
        }
      },
      "delete": {
        "tags": ["Projects"],
        "summary": "Delete project",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Project deleted"
          }
        }
      }
    },
    "/api/device-pools": {
      "get": {
        "tags": ["Device Pools"],
        "summary": "List device pools",
        "parameters": [
          {
            "name": "projectArn",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of device pools"
          }
        }
      },
      "post": {
        "tags": ["Device Pools"],
        "summary": "Create device pool",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "projectArn": {
                    "type": "string"
                  },
                  "name": {
                    "type": "string"
                  },
                  "rules": {
                    "type": "array"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Device pool created"
          }
        }
      }
    },
    "/api/devices": {
      "get": {
        "tags": ["Devices"],
        "summary": "List available devices",
        "description": "Get all available devices in AWS Device Farm",
        "responses": {
          "200": {
            "description": "List of devices"
          }
        }
      }
    },
    "/api/apps": {
      "get": {
        "tags": ["Apps"],
        "summary": "List uploaded apps",
        "parameters": [
          {
            "name": "projectId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of apps"
          }
        }
      },
      "post": {
        "tags": ["Apps"],
        "summary": "Upload app",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadAppInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "App uploaded"
          }
        }
      }
    },
    "/api/runs": {
      "get": {
        "tags": ["Runs"],
        "summary": "List test runs",
        "parameters": [
          {
            "name": "projectId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of runs"
          }
        }
      },
      "post": {
        "tags": ["Runs"],
        "summary": "Schedule test run",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRunInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Run scheduled"
          }
        }
      }
    },
    "/api/runs/{id}/stop": {
      "post": {
        "tags": ["Runs"],
        "summary": "Stop running test",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Run stopped"
          }
        }
      }
    },
    "/api/runs/{id}/artifacts": {
      "get": {
        "tags": ["Artifacts"],
        "summary": "Get run artifacts",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Run artifacts"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Project": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "awsProjectArn": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CreateProjectInput": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "description": {
            "type": "string",
            "maxLength": 500
          },
          "awsProjectArn": {
            "type": "string"
          }
        }
      },
      "UpdateProjectInput": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "awsProjectArn": {
            "type": "string"
          }
        }
      },
      "UploadAppInput": {
        "type": "object",
        "required": ["projectId", "name", "appType"],
        "properties": {
          "projectId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "appType": {
            "type": "string",
            "enum": ["android", "ios"]
          },
          "fileName": {
            "type": "string"
          },
          "fileSize": {
            "type": "number"
          }
        }
      },
      "CreateRunInput": {
        "type": "object",
        "required": ["projectId", "name"],
        "properties": {
          "projectId": {
            "type": "string"
          },
          "appId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "devicePoolArn": {
            "type": "string"
          }
        }
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Unauthorized",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    }
  }
}